+++ include/cg_local.h
@@ -208,6 +208,7 @@ char const* CG_ConfigString(int32_t index);
 //
 // cg_view.c
 //
+void init_view(void);
 void CG_DrawActiveFrame(int32_t serverTime, stereoFrame_t stereoView, qboolean demoPlayback);
 
 //